updated gitignore
