+++ setup.py
@@ -20,7 +20,9 @@ setup(
     tests_require=['pytest'],
     entry_points={
         'console_scripts': [
-            'test_run = dasc_robot_py.test_run:main',
+            'ex_rover_test = dasc_robot_py.ex_rover_test:main',
+            'ex_drone_hover = dasc_robot_py.ex_drone_hover:main',
+            'ex_rover_test_classes = dasc_robot_py.ex_rover_test_classes:main',
         ],
     },
 )
